requirements.txt
argparse==1.4.0
fake-bpy-module-3.0
imageio==2.17.0
imutils==0.5.4
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.4.60
pandas==1.4.2
piexif==1.1.3
Pillow==9.1.0
plotly==5.7.0
progressbar==2.5
PyYAML==6.0
rich==12.2.0
scipy==1.8.0
sympy==1.10.1
tqdm==4.64.0
vlips==1.3.0
